Extracted fiat related methods into FiatInfoProvider
