+++ src/providers/mod.rs
@@ -1,3 +1,4 @@
 pub mod address_info;
 pub mod ext;
+pub mod fiat;
 pub mod info;
