+++ src/services/balances.rs
@@ -2,6 +2,7 @@ use crate::cache::cache_operations::RequestCached;
 use crate::config::{balances_cache_duration, balances_request_timeout};
 use crate::models::backend::balances::Balance as BalanceDto;
 use crate::models::service::balances::{Balance, Balances};
+use crate::providers::fiat::FiatInfoProvider;
 use crate::providers::info::{DefaultInfoProvider, InfoProvider};
 use crate::utils::context::Context;
 use crate::utils::errors::ApiResult;
@@ -16,6 +17,7 @@ pub async fn balances(
     exclude_spam: bool,
 ) -> ApiResult<Balances> {
     let info_provider = DefaultInfoProvider::new(chain_id, &context);
+    let fiat_info_provider = FiatInfoProvider::new(&context);
     let url = core_uri!(
         info_provider,
         "/v1/safes/{}/balances/usd/?trusted={}&exclude_spam={}",
@@ -31,7 +33,10 @@ pub async fn balances(
         .await?;
     let backend_balances: Vec<BalanceDto> = serde_json::from_str(&body)?;
 
-    let usd_to_fiat = info_provider.exchange_usd_to(fiat).await.unwrap_or(0.0);
+    let usd_to_fiat = fiat_info_provider
+        .exchange_usd_to(fiat)
+        .await
+        .unwrap_or(0.0);
 
     let mut total_fiat = 0.0;
 
@@ -58,7 +63,7 @@ pub async fn balances(
 }
 
 pub async fn fiat_codes(context: &Context<'_>) -> ApiResult<Vec<String>> {
-    let info_provider = DefaultInfoProvider::new_no_net(&context);
+    let info_provider = FiatInfoProvider::new(&context);
     let mut fiat_codes = info_provider.available_currency_codes().await?;
 
     let usd_index = fiat_codes.iter().position(|it| it.eq("USD")).unwrap();
